Added log4net and Microsoft.Expression.Blend.SDK.WPF.4.0.0 as nuget packages.
This will re-download the packages during build.  This was setup this way due to Codeplex not allowing you to commit the packages folder created by nuget.  See http://nuget.codeplex.com/workitem/165 for more information.  The packages folder is ignored in SVN to remove the possibilities of getting the 500 error while committing to Codeplex.